app/requirements.txt
RPi.GPIO==0.7.0
adafruit-circuitpython-dht==3.7.8
